🧹 refactor: move inline components to separate files
This commit moves WordsContainer, GeneratedWords, and CountdownTimer
components from App.tsx to dedicated files in src/components/.
This improves code organization and readability of the main App component.
