+++ src/components/GeneratedWords.tsx
@@ -1,0 +1,5 @@
+const GeneratedWords = ({words} : {words: string}) => {
+  return <div className=" text-slate-500">{words}</div>
+}
+
+export default GeneratedWords;
